+++ packages/mp/src/mp/dev_env/__init__.py
@@ -1,0 +1,1 @@
+from .cli import app as app
